+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ##### Compat:
 Added user messages for when Jetpacks and Jet Boots are disabled in Heavy Air.
 
+Block Blood Magic's Air Sigil and Reliquary's Rending Gale use when in Heavy Air block.
+
+(Forge): Fixed bug where PneumaticCraft compat would put cooldown on any boots equipped.
+
 
 ### **(V.7.0.10 Changes) (1.20.1 Minecraft)**